ui(deleteuser): dedup
